pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kubepulse"
version = "1.0.0"
description = "A 3-in-1 Heartbeat Monitor for Kubernetes Manifests"
dependencies = [
    "ruamel.yaml",
]

[project.scripts]
kubepulse = "kubepulse.main:run"
